workflows to pass only accepted args to commands id:2857006
Updated generator to produce workflow handlers that interleave calls
to copy-opts in between each command in the sequence. This uses the
new feature of copy-opts that filters the arg string to just the options
supported in the called command. 

Now uses the "ignoremalformed" attribute for context 
and "resolvemodule" in command elements to make workflows
not generated for static-only to work in module contexts.


git-svn-id: file:///Users/noahcampbell/Source/ctrep/branches/controltier-3-4-support@9460 d4a9e8c4-ac4d-4c38-8d82-82c8b6c5cbc9